Added a cursed fix for f64 sometimes being represented as a u64

+++ lldf-bind/src/bind/mod.rs
@@ -5,7 +5,7 @@ use crate::core::mem::transmute_unchecked;
 /// A value returned by a DiamondFire operation.
 #[derive(Clone, Copy)]
 pub struct DFOpaqueValue {
-    _opaque_type : u8
+    _opaque_type : u64
 }
 
 unsafe impl DFValue for DFOpaqueValue {
