+++ lldf-bind/src/types/value/particle.rs
@@ -5,7 +5,7 @@ use crate::core::mem::transmute_unchecked;
 
 /// A particle effect with customisable parameters.
 pub struct Particle {
-    _opaque_type : u8
+    _opaque_type : u64
 }
 
 impl Clone for Particle {
